pyproject.toml
[project]
name = "sentinel-soil"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pyproj>=3.7.2",
    "pyyaml>=6.0.3",
    "rasterio>=1.4.4",
    "sentinelhub>=3.11.3",
    "ydata-profiling>=4.18.0",
]
